Move UI, external command and display time helpers into separate modules
