+++ version.c
@@ -1,1 +1,1 @@
-char *sklaff_version = "1.33-beta13(#6)";
+char *sklaff_version = "1.33-beta13(#9)";
